Support disabling SDN controller sync
For flat networks with bare metal compute nodes it is possible to use
the mlnx_sdn_assist mechanism driver to bind ports, without configuring
the network via SDN controller/Subnet Manager.

This change allows the SDN controller sync mechanism to be disabled if
the '[sdn] url' configuration option is not set.

Change-Id: I60a770c690d83c87636c026dbfb6c7b03a1894e7