Support disabling SDN controller sync
For flat networks with bare metal compute nodes it is possible to use
the mlnx_sdn_assist mechanism driver to bind ports, without configuring
the network via SDN controller/Subnet Manager.

This change adds a configuration option, '[sdn] enabled', and disables
the SDN controller sync mechanism if it is False.

Change-Id: I3b30ad699886bd1fd4c5cb7ed85257a3205b3aae