Fix exception handling in journal thread
The SDN journal thread can raise exceptions if requests to NEO fail, but
currently the logic used in the exception handler is not correct, and
does not catch the exception. This results in an error such as the
following:

Error on run_sync_thread: SDNConnectionError: Failed to send request 404
Client Error: NOT FOUND for url: http://1.2.3.4/neo/cloudx/Port

The result of this is that the object remains in the 'processing' state,
and the retry count is never incremented.

This change fixes the exception handling.

Change-Id: Ia15a34628905d039ee8abbeb571e71c8b83992a4
Closes-Bug: #1797105
(cherry picked from commit b9ffdc3d1e6127b19a17d8c03acf7921eccfdf29)