Set port status to ACTIVE when binding a port
Today, SDN mechanism driver does not set port status
to ACTIVE when binding on a segment.
While it has no adverse effect, it is generally a good
practice to reflect that in the port status as we have
successfuly binding the port as depicted in Neutron API ref.

"You can use the status attribute with the network and port resources to
determine whether the OpenStack Networking plug-in has successfully
completed the configuration of the resource."

Change-Id: Ie4ea52d3c0ed8ec8a9fce0871d799fc129071529