Fix deadlock of read and write lock taken by the same thread
NetworkCache get() method, in case of a cache miss called
put() method with the newly fetched network object via rpc.

Since SafeNetworkCache simply wrapped the calls with a rw lock,
it caused a dead lock in case of a cache miss since a get operation
would also put.

Proposed fix is to split the get and put logic of NetworkCache.get
into two and re-implement get in SafeNetworkCache.

Unit-test which uses a real lock in case of cache miss was
added.

Change-Id: I94fe2bcea80bb32d2c8ff8ed67194e860fd93ea9