Add Support for performing net device operations via pyroute2
This change transitions the following operations to be
executed by pyroute2:

1. set VF Admin state
2. set net device link state
3. set GUID

Today, eswitchd uses sysfs extensively to perform much
of the operations.
This change aims to reduce the use sysfs and performs
these operations in a more linux, pythonic way by
using netlink and utilizing pyroute2 library

Elevation of privleges are done by privsep which is
a safer model to be used than the existing rootwrap.

- Introduce netdev_ops abstraction - netdev_opts
- Implement abstraction using pyroute2
- Use privsep to elevate privleges to perform operations
- extend device_db to contain pf net device
- Use netdev_ops in eswitch_handler
  - When setting GUID, fallback to old sysfs functionality
    in case of failure

Change-Id: I807500e57b156fbaad85f42d70e9636c55beaa3d