Reduce agent dependency on mac address
This change reduces the agent's dependency on mac address
when actin on port CRUD operations.

a minimal dependency is still kept since we rely on MAC
address to retrieve information from neutron server
(get_device_details RPC).

The following commit makes extensive changes to eswitchd
amongst the changes, get_vnics() response dict is now keyed
by PCI address and is reflected in this commit.

Due to the extent of the changes the two commits should be
merged in one pass and is partitioned here for easier review.

Change-Id: I762a6e851c8abac92fd74faddaa5700347c8352c