Remove nested DB session from journal
We are hitting errors like this in the logs:

  ERROR oslo.service.backend.eventlet.loopingcall
  AttributeError: 'method' object has
  no attribute '_enginefacade_context'

This fixes that, by removing the nested DB sessions.
But I am not sure on what the intent was in the past, this
certainly seems simpler. Although that session param
probably needs a rename.

Change-Id: I53be4fc554e60f1c6291968110fffa1e4bafff96
(cherry picked from commit 0ec54e213946773b124bfd9f40d6ff6689a40c14)